+++ test/AlgebraTests/SparseMatrixCSR.jl
@@ -9,7 +9,7 @@ module SparseMatrixCSRTests
     maxrows=5
     maxcols=5
     maxrowsorcols=7
-    int_types=(Int32,Int64)
+    int_types=(Int32,Int)
     float_types=(Float32,Float64)
     Bi_types=(0,1)
